[pigeon] revert 3b5ae0070738103a80bb9ac12440dd69d598ad2c (#2987)
* start

* Class file transfer and skeleton comments

* relocate generator classes and update imports

* merge main and changelog update

* update openSink to openWriteSink

* Revert "[pigeon] Relocates generator classes and updates imports (#2985)"

This reverts commit 3b5ae0070738103a80bb9ac12440dd69d598ad2c.
