+++ packages/pigeon/lib/generator_tools.dart
@@ -9,7 +9,7 @@ import 'dart:mirrors';
 import 'ast.dart';
 
 /// The current version of pigeon. This must match the version in pubspec.yaml.
-const String pigeonVersion = '4.2.15';
+const String pigeonVersion = '4.2.14';
 
 /// Read all the content from [stdin] to a String.
 String readStdin() {
